+++ src/node/config.ts
@@ -31,14 +31,21 @@ export interface SharedConfig {
    */
   root?: string
   /**
-   * Import alias. Can only be exact mapping, does not support wildcard syntax.
+   * Import alias. The entries can either be exact request -> request mappings
+   * (exact, no wildcard syntax), or request path -> fs directory mappings.
+   * When using directory mappings, the key **must start and end with a slash**.
    *
    * Example `vite.config.js`:
    * ``` js
    * module.exports = {
    *   alias: {
+   *     // alias package names
    *     'react': '@pika/react',
    *     'react-dom': '@pika/react-dom'
+   *
+   *     // alias a path to a fs directory
+   *     // the key must start and end with a slash
+   *     '/@foo/': path.resolve(__dirname, 'some-special-dir')
    *   }
    * }
    * ```
@@ -187,21 +194,13 @@ export interface BuildConfig extends SharedConfig {
    */
   ssr?: boolean
 
-  // The following are API only and not documented in the CLI. -----------------
+  // The following are API / config only and not documented in the CLI. --------
   /**
    * Will be passed to rollup.rollup()
    *
    * https://rollupjs.org/guide/en/#big-list-of-options
    */
   rollupInputOptions?: RollupInputOptions
-  /**
-   * Will be passed to @rollup/plugin-commonjs
-   * https://github.com/rollup/plugins/tree/commonjs-v11.1.0/packages/commonjs#namedexports
-   * This config can be removed after master branch is released.
-   * But there are some issues blocking it:
-   * https://github.com/rollup/plugins/issues/392
-   */
-  rollupPluginCommonJSNamedExports?: Record<string, string[]>
   /**
    * Will be passed to bundle.generate()
    *
@@ -214,6 +213,11 @@ export interface BuildConfig extends SharedConfig {
    * https://github.com/vuejs/rollup-plugin-vue/blob/next/src/index.ts
    */
   rollupPluginVueOptions?: Partial<RollupPluginVueOptions>
+  /**
+   * Will be passed to @rollup/plugin-node-resolve
+   * https://github.com/rollup/plugins/tree/master/packages/node-resolve#dedupe
+   */
+  rollupDedupe?: string[]
   /**
    * Whether to log asset info to console
    * @default false
@@ -291,7 +295,10 @@ export async function resolveConfig(
   }
 
   if (!resolvedPath) {
-    return
+    // load environment variables
+    return {
+      env: loadEnv(mode, cwd)
+    }
   }
 
   try {
@@ -313,14 +320,14 @@ export async function resolveConfig(
       // 2. if we reach here, the file is ts or using es import syntax.
       // transpile es import syntax to require syntax using rollup.
       const rollup = require('rollup') as typeof Rollup
-      const esbuilPlugin = await createEsbuildPlugin(false, {})
+      const esbuildPlugin = await createEsbuildPlugin(false, {})
       const bundle = await rollup.rollup({
         external: (id: string) =>
           (id[0] !== '.' && !path.isAbsolute(id)) ||
           id.slice(-5, id.length) === '.json',
         input: resolvedPath,
         treeshake: false,
-        plugins: [esbuilPlugin]
+        plugins: [esbuildPlugin]
       })
 
       const {
@@ -345,11 +352,7 @@ export async function resolveConfig(
       }
     }
 
-    // load environment variables
-    const env = loadEnv(mode, config.root || cwd)
-    debug(`env: %O`, env)
-    config.env = env
-
+    config.env = loadEnv(mode, config.root || cwd)
     debug(`config resolved in ${Date.now() - start}ms`)
 
     config.__path = resolvedPath
@@ -439,5 +442,6 @@ function loadEnv(mode: string, root: string): Record<string, string> {
     }
   }
 
+  debug(`env: %O`, env)
   return env
 }