+++ src/node/server/serverPluginModuleRewrite.ts
@@ -19,7 +19,8 @@ import {
   ensureMapEntry,
   rewriteFileWithHMR,
   hmrClientPublicPath,
-  hmrDirtyFilesMap
+  hmrDirtyFilesMap,
+  latestVersionsMap
 } from './serverPluginHmr'
 import {
   readBody,
@@ -28,7 +29,6 @@ import {
   resolveRelativeRequest
 } from '../utils'
 import chalk from 'chalk'
-import slash from 'slash'
 import { moduleRE } from './serverPluginModuleResolve'
 
 const debug = require('debug')('vite:rewrite')
@@ -227,8 +227,6 @@ export function rewriteImports(
 }
 
 const bareImportRE = /^[^\/\.]/
-const indexRE = /\/index\.\w+$/
-const indexRemoveRE = /\/index(\.\w+)?$/
 
 export const resolveImport = (
   root: string,
@@ -254,14 +252,12 @@ export const resolveImport = (
     }
 
     // 3. resolve extensions.
-    const file = slash(resolver.requestToFile(pathname))
-    const resolvedExt = path.extname(file)
-    if (resolvedExt !== path.extname(pathname)) {
-      const indexMatch = file.match(indexRE)
-      if (indexMatch) {
-        pathname = pathname.replace(indexRemoveRE, '') + indexMatch[0]
-      } else {
-        pathname += resolvedExt
+    const ext = resolver.resolveExt(pathname)
+    if (ext) {
+      pathname += ext
+      if (ext[0] === '/') {
+        // in aliased cases the inferred ext can contain multiple slashes
+        pathname = pathname.replace(/\/\/+/g, '/')
       }
     }
 
@@ -273,14 +269,13 @@ export const resolveImport = (
     // 5. force re-fetch dirty imports by appending timestamp
     if (timestamp) {
       const dirtyFiles = hmrDirtyFilesMap.get(timestamp)
-      // only force re-fetch if this is a marked dirty file (in the import
-      // chain of the changed file) or a vue part request (made by a dirty
-      // vue main request)
-      if (
-        (dirtyFiles && dirtyFiles.has(pathname)) ||
-        /\?type=(template|style)/.test(id)
-      ) {
+      // only rewrite if:
+      if (dirtyFiles && dirtyFiles.has(pathname)) {
+        // 1. this is a marked dirty file (in the import chain of the changed file)
         query += `${query ? `&` : `?`}t=${timestamp}`
+      } else if (latestVersionsMap.has(pathname)) {
+        // 2. this file was previously hot-updated and has an updated version
+        query += `${query ? `&` : `?`}t=${latestVersionsMap.get(pathname)}`
       }
     }
     return pathname + query